refactor(cli): move programs to modules
