--- apps/cli/bash.ts
@@ -1,1 +1,0 @@
-export { BASH_SOURCE } from "../../core/fs/bin";
