--- apps/cli/init.ts
@@ -1,1 +1,0 @@
-export { INIT_SOURCE } from "../../core/fs/bin";
