--- apps/cli/kill.ts
@@ -1,1 +1,0 @@
-export { KILL_SOURCE } from "../../core/fs/bin";
