--- apps/cli/login.ts
@@ -1,1 +1,0 @@
-export { LOGIN_SOURCE } from "../../core/fs/bin";
