--- apps/cli/mkdir.ts
@@ -1,1 +1,0 @@
-export { MKDIR_SOURCE } from "../../core/fs/bin";
