--- apps/cli/mv.ts
@@ -1,1 +1,0 @@
-export { MV_SOURCE } from "../../core/fs/bin";
