--- apps/cli/nano.ts
@@ -1,1 +1,0 @@
-export { NANO_SOURCE } from "../../core/fs/bin";
