--- apps/cli/ping.ts
@@ -1,1 +1,0 @@
-export { PING_SOURCE } from "../../core/fs/bin";
