+++ apps/cli/programs/reboot.ts
@@ -1,0 +1,6 @@
+import type { SyscallDispatcher } from "../../core/kernel/syscalls";
+
+export async function main(syscall: SyscallDispatcher): Promise<number> {
+    await syscall('reboot');
+    return 0;
+}
