--- apps/cli/ps.ts
@@ -1,1 +1,0 @@
-export { PS_SOURCE } from "../../core/fs/bin";
