--- apps/cli/snapshot.ts
@@ -1,1 +1,0 @@
-export { SNAPSHOT_SOURCE } from "../../core/fs/bin";
