--- apps/cli/src/reboot.ts
@@ -1,6 +1,0 @@
-export const REBOOT_SOURCE = `
-  async (syscall) => {
-    await syscall('reboot');
-    return 0;
-  }
-`;
