--- apps/cli/ulimit.ts
@@ -1,1 +1,0 @@
-export { ULIMIT_SOURCE } from "../../core/fs/bin";
