+++ apps/index.ts
@@ -3,18 +3,18 @@ export * from "./examples/sshClient";
 export * from "./examples/desktop";
 export * from "./examples/webBrowser";
 
-export * from "./cli/bash";
-export * from "./cli/init";
-export * from "./cli/kill";
-export * from "./cli/login";
-export * from "./cli/ls";
-export * from "./cli/mkdir";
-export * from "./cli/rm";
-export * from "./cli/mv";
-export * from "./cli/nano";
-export * from "./cli/ping";
-export * from "./cli/ps";
-export * from "./cli/snapshot";
-export * from "./cli/ulimit";
+export * from "./cli/programs/bash";
+export * from "./cli/programs/init";
+export * from "./cli/programs/kill";
+export * from "./cli/programs/login";
+export * from "./cli/programs/ls";
+export * from "./cli/programs/mkdir";
+export * from "./cli/programs/rm";
+export * from "./cli/programs/mv";
+export * from "./cli/programs/nano";
+export * from "./cli/programs/ping";
+export * from "./cli/programs/ps";
+export * from "./cli/programs/snapshot";
+export * from "./cli/programs/ulimit";
 
 export { BUNDLED_APPS } from "../core/fs/bin";
